Vaporcredits balance is now limited to integer values

+++ src/migrations/20260818_create_wallets_table.js
@@ -5,7 +5,7 @@ export async function up() {
     CREATE TABLE IF NOT EXISTS vapor_credits_wallets (
       id UUID PRIMARY KEY,
       user_id UUID NOT NULL UNIQUE REFERENCES users(id),
-      balance NUMERIC(12, 2) NOT NULL DEFAULT 0 CHECK (balance >= 0),
+      balance INTEGER NOT NULL DEFAULT 0 CHECK (balance >= 0),
       updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
     )
   `);
